Update documentation examples with datasets module
Since the examples in our documentation for the Manifold and RFECV
visualizers take an exceptionally long time to run, it was decided not
to include them in the new plot directive #687. Additionally, since the
JointPlot visualizer is currently undergoing some changes, we'll wait
until those are complete to use the plot directive in its docs. However,
this PR does update the code examples for those visualizers with the new
datasets module and makes some minor formatting changes to help make our
documentation more consistent.

See also: #926