Extend DataVisualizer and Update DataVisualizer Subclasses (PR #927)
This PR extends our DataVisualize class in order to provide a consistent method for it's subclasses to identify classes, assign colors and provide test coverage. This PR closes #923, and makes updates to a number of visualizers, including RadViz and ParallelCoords. As some color params are being deprecated/consolidated, this PR also updates the FAQ on how to pass custom colors to a visualizer.