Check if an estimator is already fitted before calling fit (#925)
Added two helper utilities: `is_fitted` and `check_fitted` that control the fitted estimator checking. If the user supplies `is_fitted='auto'`, the default, the model visualizer will check if it's fitted using a mechanism recommended by the scikit-learn team and Stack Overflow and will not fit a fitted estimator. Otherwise, the model visualizer will accept the recommendation of the user to fit or not fit the wrapped estimator. 

Fixes #297