Do not install tests subpackages with pip
The new module rules mean that directories without a __init__.py
file can be found, which meant that the find_packages method was
discovering tests subpackages. By adding the wildcard flag to
excludes, we omit these packages from our builds and they will
not be installed by pip or setup.py on the next release.

Fixes #996