Update kneed algorithm (#1121)
Updated the kneed algorithm to the latest version of the port from the original author. This was not a 100% update and extra features like argument validation and polynomial fit were removed since they are not directly applicable to Yellowbrick's use case. This PR will reduce unnecessary warnings and surprises where there is a "clear" knee but the algorithm was previously unable to find it.