Housekeeping
This PR fixes #1113 by handling some housekeeping issues. We've changed
our master branch to the main branch and created a release issue
template (though I don't think it is used automatically by GitHub).
We've also checked the new pip resolver and all of our dependencies are
compatibile. This PR also updates our CI to test Python 3.8 and 3.9 to
ensure we're testing on the latest version of Python.