FIXES 1086: Corrects legend issues other than R2 in PredictionError (#1212)
* BUG: Corrects legend issues other than R2 in PredictionError

This PR fixes issue #1086

Unfortunately, the generic .score() method in scikit-learn does not return the name of the scoring metrics used. The fix I have patched will return the correct label if and when the estimator has an attribute providing the name of the scoring metric otherwise it will fall back to the default value i.e. R2

Co-authored-by: Pkaf <prrajkafle@gmail.com>
Co-authored-by: Gray <lwgray@gmail.com>