Random input feature dropping curve, model selection visualization [issue #1024] (#1206)
* Create dropping_curve.py

* Add imports

* Add a stub DroppingCurve class

Only has API documentation for now

* Initialize the class with the __init__ variables

* Rename from test set to validation set

* Outline for the fit() method - need to figure out how to wrap the feature subsetting though

* Adapt draw() method from learning_curve.py

* Add finalize method

* Add helper quick method for dropping_curve

* Add random scoring method and also the start of a feature-dropping curve. Need to solidify the feature_sizes_ variable though

Perhaps could use Percentile best, too

* Convert feature_sizes_ to integers that we can use for plotting later

* Fix some typos

* Quick-script example for dropping_curve

Seems to create a buch of empty figures - not sure why?

* 2022 Maintenance (#1207)

This PR switches to a YB-specific data lake for datasets storage and updates the prior nltk dependency that has a CVE.

* FIXES 1086: Corrects legend issues other than R2 in PredictionError (#1212)

* BUG: Corrects legend issues other than R2 in PredictionError

This PR fixes issue #1086

Unfortunately, the generic .score() method in scikit-learn does not return the name of the scoring metrics used. The fix I have patched will return the correct label if and when the estimator has an attribute providing the name of the scoring metric otherwise it will fall back to the default value i.e. R2

Co-authored-by: Pkaf <prrajkafle@gmail.com>
Co-authored-by: Gray <lwgray@gmail.com>

* BUG: Fixes axes limit for PredictionError plot #1193 (#1208)

Co-authored-by: Pkaf <prrajkafle@gmail.com>

* Updates sklearn  to v1.0.0 (#1217)

* PrePredict Estimator (#1189)

* BUG: Adds missing X and Y axes labels in ClassificationReport (#1210)

Co-authored-by: Pkaf <prrajkafle@gmail.com>
Co-authored-by: Larry Gray <lwgray@gmail.com>

* version bump v1.4

* Fixed is_fitted parameter not setting to given value (#1221)

Set `super().init()` for the visualizer to include `is_fitted`

* Create tests for `is_fitted` parameter for ResidualsPlot and PredictionError (#1223)

This PR adds tests to #1221. These tests assure us that the is_fitted param's state is maintained when the visualizer is instantiated.

I have made the following changes:

Added test for ResidualPlots
Added test for PredictionError
Removed unnecessary import so that flake8 passed

* Remove try/except clauses from tests (#1224)

This PR fixes issue #1218 which reported technical debt in which there were several try/except clauses within tests that needed to be removed. This was first identified by @rebeccabilbro in PR #1197

* Create tests

* Modify dropping_curve.py to be compliant with Sklearn version bump

The switch to Sklearn 1.0 lead to a deep copy error of ax within the sk_validation_curve call.
@bbengfort and I fixed this by removing the input of self into the the score_func of SelectKBest
and replaced it with a lambda function

* Add Dropping Curve Test

* Fix quick_method and classifier test

* Address review comments to turn asserts to YellowbrickValueError, remove clustering from test, and rearrange parameters so ax=none is in the second position

* Fix YellowbrickValueError Errors

Co-authored-by: Benjamin Bengfort <benjamin@bengfort.com>
Co-authored-by: pkaf <pkaf@users.noreply.github.com>
Co-authored-by: Pkaf <prrajkafle@gmail.com>
Co-authored-by: Gray <lwgray@gmail.com>
Co-authored-by: Lawrence Gray <lwgray@gmail.com>